Add Explosive Turret

+++ Assets/ScriptableObjects/Data/Turret Types/InstantTurretSO.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Instant Turret", menuName = "Create Turret/Instant Turret")]
 public class InstantTurretSO : BaseTurretSO
 {
-    public int explosiveDamage;
-    public int castTime;
-    public int range;
+    public float explosiveDamage;
+    public float castTime;
+    public float range;
 }
